requirements.txt
ccxt
pandas
python-dotenv
ta
pytest>=7.0